feat(contact-controller): improving delete method

+++ src/main/java/p2p/restapi/controller/ContactController.java
@@ -3,10 +3,12 @@ package p2p.restapi.controller;
 import java.util.List;
 import java.util.Optional;
 
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import p2p.restapi.ContactNotFoundException;
@@ -37,7 +39,10 @@ public class ContactController {
     }
 
     @PostMapping
-    public ResponseEntity<Contact> save(@RequestBody Contact contact) {
+    public ResponseEntity<?> save(@RequestBody @Valid Contact contact, BindingResult result) {
+        if (result.hasErrors()) {
+            return new ResponseEntity<>(result.getAllErrors(), HttpStatus.BAD_REQUEST);
+        }
         try {
             Contact savedContact = repository.save(contact);
             return new ResponseEntity<>(savedContact, HttpStatus.CREATED);
@@ -45,6 +50,7 @@ public class ContactController {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
+
 
 
     @PutMapping
@@ -62,14 +68,17 @@ public class ContactController {
     @DeleteMapping("/{id}")
     public ResponseEntity<String> delete(@PathVariable Long id) {
         try {
-            repository.deleteById(id);
-            return new ResponseEntity<>("Contato excluído com sucesso", HttpStatus.OK);
-        } catch (EmptyResultDataAccessException e) {
-            return new ResponseEntity<>("Contato não encontrado com o ID: " + id, HttpStatus.NOT_FOUND);
+            if (repository.existsById(id)) {
+                repository.deleteById(id);
+                return new ResponseEntity<>("Contato excluído com sucesso", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Contato não encontrado com o ID: " + id, HttpStatus.NOT_FOUND);
+            }
         } catch (Exception e) {
             return new ResponseEntity<>("Erro ao excluir o contato", HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
+
